+++ test/plugins/test_zero.py
@@ -256,7 +256,8 @@ class ZeroPluginTest(IOMixin, PluginTestCase):
 
         mf = MediaFile(syspath(item.path))
         assert mf.comments is None
-        assert mf.disc == 0
+        assert mf.disc is None
+        assert mf.disctotal is None
 
     def test_omit_single_disc_with_tags_multi(self):
         item = self.add_item_fixture(
@@ -271,6 +272,7 @@ class ZeroPluginTest(IOMixin, PluginTestCase):
         mf = MediaFile(syspath(item.path))
         assert mf.comments is None
         assert mf.disc == 1
+        assert mf.disctotal == 4
 
     def test_omit_single_disc_only_change_single(self):
         item = self.add_item_fixture(disctotal=1, disc=1)
@@ -280,7 +282,8 @@ class ZeroPluginTest(IOMixin, PluginTestCase):
             item.write()
 
         mf = MediaFile(syspath(item.path))
-        assert mf.disc == 0
+        assert mf.disc is None
+        assert mf.disctotal is None
 
     def test_omit_single_disc_only_change_multi(self):
         item = self.add_item_fixture(disctotal=4, disc=1)
@@ -291,6 +294,7 @@ class ZeroPluginTest(IOMixin, PluginTestCase):
 
         mf = MediaFile(syspath(item.path))
         assert mf.disc == 1
+        assert mf.disctotal == 4
 
     def test_empty_query_n_response_no_changes(self):
         item = self.add_item_fixture(